Add function that populates screen and operands
Functions are activated when buttons are pressed.
They have a similar structure, as both of them
depend on the ID of the pressed button to define
the command inputed.
Still sorting out how to make the operands
function work properly, as it does not respond
correctly to multiple operations in sequence.